refactor: 重构项目结构使符合作业要求
1. 重构项目结构
2. 解决路径问题
