--- 221801107/.eslintignore
@@ -1,2 +1,0 @@
-# 忽略 dist 目录语法检查
-dist/*
