--- 221801107/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "221801107",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "private": true,
-  "dependencies": {
-    "jest": "^26.6.3"
-  },
-  "scripts": {
-    "start": "node src/index.js input.txt output.txt",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "eslint": "7.2.0",
-    "eslint-config-airbnb": "18.2.1",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.21.5",
-    "eslint-plugin-react-hooks": "1.7.0"
-  }
-}
