--- 221801107/src/config/paths.js
@@ -1,7 +1,0 @@
-const path = require("path");
-
-const ROOT_PATH = path.join(__dirname, "..", "..");
-
-module.exports = {
-  ROOT_PATH,
-};
