+++ 221801107/src/filterChinese.js
@@ -1,0 +1,5 @@
+const { CHINESE_REGEX } = require("./regex");
+
+const filterChinese = (content) => content.replace(new RegExp(CHINESE_REGEX, "gm"), "");
+
+module.exports = filterChinese;
