+++ 221801107/src/heap.js
@@ -1,5 +1,3 @@
-/* eslint-disable no-param-reassign */
-/* eslint-disable class-methods-use-this */
 class Heap {
   /**
    * 堆构造函数
