--- 221801107/src/lib/wordCount/constants/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  ...require('./regex'),
-};
