--- 221801107/src/lib/wordCount/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  ...require('./character'),
-  ...require('./row'),
-  ...require('./word'),
-};
