--- 221801107/src/lib/wordCount/utils/filterChinese.js
@@ -1,5 +1,0 @@
-const { CHINESE_REGEX } = require("../constants");
-
-const filterChinese = (content) => content.replace(new RegExp(CHINESE_REGEX, 'gm'), "");
-
-module.exports = filterChinese;
